requirements.txt
numpy<2
requests~=2.32.3
dataclasses~=0.6
dacite~=1.8.1
grpcio>=1.66.1
grpcio-tools>=1.66.1
protobuf
grpcio-reflection~=1.66.1
grpcio-health-checking~=1.66.1
pandas
openpyxl
scikit-learn
langchain-community
faiss-cpu
gigachat
tiktoken
setuptools~=68.2.2
aiocache~=0.12.3